+++ src/Framework/Logger/LoggerInterface.php
@@ -114,7 +114,27 @@ interface LoggerInterface
     public function fetchAll();
 
     /**
+     * @return array
+     */
+    public function fetchLatest($limit = 100);
+
+    /**
+     * @return int
+     */
+    public function getTotalLogs();
+
+    /**
+     * @return int
+     */
+    public function getLogSizeInBytes();
+
+    /**
      * @return void
      */
     public function dropDatabaseTable();
+
+    /**
+     * @return bool
+     */
+    public function isDownloadLogRequested();
 }