+++ src/Views/menu-debug-log.php
@@ -3,13 +3,14 @@
 use PublishPress\Future\Modules\Settings\HooksAbstract;
 use PublishPress\Future\Core\DI\Container;
 use PublishPress\Future\Core\DI\ServicesAbstract;
+use PublishPress\Future\Framework\Logger\LoggerInterface;
 
 $container = Container::getInstance();
 $hooks = $container->get(ServicesAbstract::HOOKS);
 
 defined('ABSPATH') or die('Direct access not allowed.');
 
-print '<p>' . esc_html__(
+echo '<p>' . esc_html__(
     'Below is a dump of the debugging table, this should be useful for troubleshooting.',
     'post-expirator'
 ) . '</p>';
@@ -19,21 +20,62 @@ $showSideBar = $hooks->applyFilters(
     ! defined('PUBLISHPRESS_FUTURE_LOADED_BY_PRO')
 );
 
-print '<div class="pp-columns-wrapper' . ($showSideBar ? ' pp-enable-sidebar' : '') . '">';
-print '<div class="pp-column-left">';
+echo '<div class="pp-columns-wrapper' . ($showSideBar ? ' pp-enable-sidebar' : '') . '">';
+echo '<div class="pp-column-left">';
 
-$debug = Container::getInstance()->get(ServicesAbstract::DEBUG);
-$results = $debug->fetchAll();
+/**
+ * @var LoggerInterface $logger
+ */
+$logger = Container::getInstance()->get(ServicesAbstract::LOGGER);
+$results = $logger->fetchLatest(500);
+$totalLogs = $logger->getTotalLogs();
+$logSizeInBytes = $logger->getLogSizeInBytes();
 
 if (empty($results)) {
-    print '<p>' . esc_html__('Debugging table is currently empty.', 'post-expirator') . '</p>';
+    echo '<p>' . esc_html__('Debugging table is currently empty.', 'post-expirator') . '</p>';
 }
 
 if (! empty($results)) {
 
-    // Add copy button
-    print '<button id="copy-debug-log" class="button">' . esc_html__('Copy Debug Log', 'post-expirator') . '</button>';
-    print '<br><br>';
+    echo '<div class="pp-debug-log">';
+    echo '<textarea readonly>';
+    foreach ($results as $result) {
+        printf("%s: %s\n", $result['timestamp'], esc_html($result['message']));
+    }
+    echo '</textarea>';
+
+    $totalDisplayedLogs = count($results);
+
+    if ($totalLogs > $totalDisplayedLogs) {
+        echo '<p id="debug-log-length">' . sprintf(
+        // translators: %s is the number of results in the debug log. %s is the size of the log in the most appropriate unit.
+        esc_html__('Showing the latest %d of %d results. The approximate size of the log is %s.', 'post-expirator'),
+        $totalDisplayedLogs,
+        $totalLogs,
+            PostExpirator_Util::formatBytes($logSizeInBytes)
+        ) . '</p>';
+    } else {
+        echo '<p id="debug-log-length">' . sprintf(
+            // translators: %s is the size of the log in the most appropriate unit.
+            esc_html__('Showing all %d results. The approximate size of the log is %s.', 'post-expirator'),
+            $totalLogs,
+            PostExpirator_Util::formatBytes($logSizeInBytes)
+        ) . '</p>';
+    }
+
+    echo '<div class="pp-debug-log-actions">';
+
+    $nonce = wp_create_nonce('publishpress_future_download_log');
+
+    echo '<button id="copy-debug-log" class="button">' . esc_html__('Copy Debug Log', 'post-expirator') . '</button>';
+
+    echo '<a href="' . esc_url(add_query_arg([
+        'action' => 'publishpress_future_debug_log',
+        'nonce' => $nonce,
+    ], admin_url('admin.php'))) . '" class="button">'
+        . esc_html__('Download Entire Log', 'post-expirator') . '</a>';
+
+    echo '</div>';
 
     // Add JavaScript to handle copying
     ?>
@@ -47,23 +89,7 @@ if (! empty($results)) {
     </script>
     <?php
 
-    print '<div class="pp-debug-log">';
-    print '<textarea readonly>';
-    foreach ($results as $result) {
-        printf("%s: %s\n", $result['timestamp'], esc_html($result['message']));
-    }
-    print '</textarea>';
-
-    $totalOfResults = count($results);
-
-    print '<p id="debug-log-length">' . sprintf(
-        // translators: %s is the number of results in the debug log. %s is the size of the log in KB.
-        esc_html__('Debug log contains %s results. The approximate size of the log is %s KB.', 'post-expirator'),
-        $totalOfResults,
-        '###'
-    ) . '</p>';
-
-    print '</div>';
+    echo '</div>';
 
     // Add JavaScript to auto-scroll textarea to the end
     ?>
@@ -71,19 +97,14 @@ if (! empty($results)) {
     document.addEventListener('DOMContentLoaded', function() {
         const debugLog = document.querySelector('.pp-debug-log textarea');
         debugLog.scrollTop = debugLog.scrollHeight;
-
-        // Measure the size in bytes of the debug log
-        const debugLogSize = debugLog.value.length / 1024;
-        document.getElementById('debug-log-length').textContent =
-            document.getElementById('debug-log-length').textContent.replace('###', debugLogSize.toFixed(2));
     });
     </script>
     <?php
 }
 
-print '</div>';
+echo '</div>';
 
 if ($showSideBar) {
     include __DIR__ . '/ad-banner-right-sidebar.php';
 }
-print '</div>';
+echo '</div>';